Add repositories configuration (#13)
* Add action util helper

Add function that allows for the retrieval of input information. Optional `trimWhitespace` and `defaultValue` can be specified.

* Add schema util helper

Add a helper function that interacts with `ajv` to validate schemas
that will be used within the application.

* Add test files for repository schema validation

The repository input will have to conform to a specified schema. This
change contains the tests that should confirm that this functionality is
working as intended.

* Add schema for repositories input

Add the schema for repositories. If the JSON input fails, or schema
validation fails, this results in undefined being returned.

* Add repository usage to main function

Replace the template `run` code to utilise the new functionality. This
also includes the addition of the input within the `action.yml`
configuration file.

* Update README